+++ CHANGELOG.md
@@ -26,6 +26,7 @@ The types of changes are:
 - Support for passing in an AC string as part of a fides string for the TCF overlay [#4308](https://github.com/ethyca/fides/pull/4308)
 - Added support for overriding the save user preferences API call with a custom fn provided through Fides.init [#4318](https://github.com/ethyca/fides/pull/4318)
 - Return AC strings in GET Privacy Experience meta and allow saving preferences against AC strings [#4295](https://github.com/ethyca/fides/pull/4295)
+- New GET Privacy Experience Meta Endpoint [#4328](https://github.com/ethyca/fides/pull/4328)
 - Access and erasure support for SparkPost [#4328](https://github.com/ethyca/fides/pull/4238)
 - Access and erasure support for Iterate [#4332](https://github.com/ethyca/fides/pull/4332)